@@ -1,30 +1,25 @@
-import { DoWork, ObservableWorker } from 'observable-webworker';
-import { Observable } from 'rxjs';
 import { range } from 'lodash';
-import {
-    map
-} from 'rxjs/operators';
+import { DoWork, runWorker } from 'observable-webworker';
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
-@ObservableWorker()
 export class PrimeNumberCounterWorker implements DoWork<number, number> {
   public work(input$: Observable<number>): Observable<number> {
-
     return input$.pipe(
-      map(numberLim => this.randomNumberGenerate(numberLim))
+      map((numberLim) => this.randomNumberGenerate(numberLim)),
     );
   }
 
   randomNumberGenerate(rangeLimit: number): number {
-
     let count = 0;
     const array = range(rangeLimit);
-    array.forEach((item, index) => {
-      this.isPrime(item) ? count += 1 : null;
+    array.forEach((item) => {
+      this.isPrime(item) ? (count += 1) : null;
     });
     return count;
   }
 
-   isPrime(num) {
+  isPrime(num) {
     if (num < 2) {
       return false;
     }
@@ -36,3 +31,4 @@ export class PrimeNumberCounterWorker implements DoWork<number, number> {
     return true;
   }
 }
+runWorker(PrimeNumberCounterWorker);
