@@ -1,11 +1,11 @@
-import {Component, ViewChild} from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.less']
+  styleUrls: ['./app.component.less'],
 })
 export class AppComponent {
   title = 'marlett-app';
-
+  constructor() {}
 }
