@@ -1,19 +1,22 @@
 import { ActionReducerMap } from '@ngrx/store';
 import * as fromWebWorkerPrimeCounterState from './reducers/webWorkerPrimeNumberCounter/web-worker-prime-number-counter.reducer';
 
-export function defaultReducer<T>(state: T) { return state; }
+export function defaultReducer<T>(state: T) {
+  return state;
+}
 
 export const initialReducerMap = {
-  webWorkerPrimeNumberCounterState : defaultReducer,
+  webWorkerPrimeNumberCounterState: defaultReducer,
 } as ActionReducerMap<ApplicationState>;
 
 export function getInitialState() {
-  let g = {
-    webWorkerPrimeNumberCounterState: fromWebWorkerPrimeCounterState .initialState,
+  const g = {
+    webWorkerPrimeNumberCounterState:
+      fromWebWorkerPrimeCounterState.initialState,
   } as ApplicationState;
   return g;
 }
 
 export interface ApplicationState {
-  webWorkerPrimeNumberCounterState: fromWebWorkerPrimeCounterState .State;
+  webWorkerPrimeNumberCounterState: fromWebWorkerPrimeCounterState.State;
 }
