@@ -1,8 +1,8 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { AboutSectionComponent } from './about-section.component';
-import {AngularFontAwesomeModule} from 'angular-font-awesome';
-import {WheelItemComponent} from '../wheel-item/wheel-item.component';
+import { AngularFontAwesomeModule } from 'angular-font-awesome';
+import { WheelItemComponent } from '../wheel-item/wheel-item.component';
 
 describe('AboutSectionComponent', () => {
   let component: AboutSectionComponent;
@@ -10,10 +10,9 @@ describe('AboutSectionComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [ AngularFontAwesomeModule ],
-      declarations: [ AboutSectionComponent, WheelItemComponent ]
-    })
-    .compileComponents();
+      imports: [AngularFontAwesomeModule],
+      declarations: [AboutSectionComponent, WheelItemComponent],
+    }).compileComponents();
   }));
 
   beforeEach(() => {
