@@ -2,8 +2,8 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { HeroSectionComponent } from '../hero-section/hero-section.component';
 import { HomeComponent } from './home.component';
 import { AboutSectionComponent } from '../about-section/about-section.component';
-import {AngularFontAwesomeModule} from 'angular-font-awesome';
-import {WheelItemComponent} from '../wheel-item/wheel-item.component';
+import { AngularFontAwesomeModule } from 'angular-font-awesome';
+import { WheelItemComponent } from '../wheel-item/wheel-item.component';
 
 describe('HomeComponent', () => {
   let component: HomeComponent;
@@ -11,11 +11,14 @@ describe('HomeComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [ AngularFontAwesomeModule ],
-      declarations: [ HomeComponent,
-        HeroSectionComponent, AboutSectionComponent, WheelItemComponent ]
-    })
-    .compileComponents();
+      imports: [AngularFontAwesomeModule],
+      declarations: [
+        HomeComponent,
+        HeroSectionComponent,
+        AboutSectionComponent,
+        WheelItemComponent,
+      ],
+    }).compileComponents();
   }));
 
   beforeEach(() => {
