@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { WheelItemComponent } from './wheel-item.component';
-import {AngularFontAwesomeModule} from 'angular-font-awesome';
+import { AngularFontAwesomeModule } from 'angular-font-awesome';
 
 describe('WheelItemComponent', () => {
   let component: WheelItemComponent;
@@ -9,16 +9,19 @@ describe('WheelItemComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [ AngularFontAwesomeModule ],
-      declarations: [ WheelItemComponent ]
-    })
-    .compileComponents();
+      imports: [AngularFontAwesomeModule],
+      declarations: [WheelItemComponent],
+    }).compileComponents();
   }));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(WheelItemComponent);
     component = fixture.componentInstance;
-    component.wheelItem = {icon: 'github', href: 'https://github.com/jmarlett93', text: 'GitHub Profile'};
+    component.wheelItem = {
+      icon: 'github',
+      href: 'https://github.com/jmarlett93',
+      text: 'GitHub Profile',
+    };
     fixture.detectChanges();
   });
 
