@@ -5,13 +5,13 @@ import { EffectsModule } from '@ngrx/effects';
 import * as fromWebWorkerPrimeNumber from './webWorkerPrimeNumberCounter/web-worker-prime-number-counter.reducer';
 import { WebWorkerPrimeNumberCounterEffects } from './webWorkerPrimeNumberCounter/web-worker-prime-number-counter.effects';
 @NgModule({
-  declarations: [],
   imports: [
-    StoreModule.forFeature('webWorkerPrimeNumberCounterState',
-      fromWebWorkerPrimeNumber.reducer),
-    EffectsModule.forFeature([WebWorkerPrimeNumberCounterEffects])
+    StoreModule.forFeature(
+      'webWorkerPrimeNumberCounterState',
+      fromWebWorkerPrimeNumber.reducer,
+    ),
+    EffectsModule.forFeature([WebWorkerPrimeNumberCounterEffects]),
   ],
   providers: [WebWorkerPrimeNumberCounterEffects],
 })
 export class WebWorkerPrimeNumberCounterStateModule {}
-
