@@ -1,4 +1,5 @@
-import {  createFeatureSelector } from '@ngrx/store';
 import { ApplicationState } from '../../app.state';
 
-export const selectWebWorkerPrimeNumberCounterState = (state: ApplicationState) => state.webWorkerPrimeNumberCounterState;
+export const selectWebWorkerPrimeNumberCounterState = (
+  state: ApplicationState,
+) => state.webWorkerPrimeNumberCounterState;
